More refactoring, clippy fixes for expression.rs
